Slim PrometheusType interface
